backend/requirements.txt
flask
Pillow
boto3
flask-cors
gunicorn

